blog/requirements.txt
minio
flask
gunicorn
python-dotenv